+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased][unreleased]
--
+
+## [1.0.3] - 2016-03-22
+- Added webhook listener.
+- Added scheduled events to check transaction status.
+- Updated gateway settings.
 
 ## [1.0.2] - 2016-03-02
 - Added get_settings function.
@@ -18,6 +22,7 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2016-02-01
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.2...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.3...HEAD
+[1.0.3]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.0...1.0.1